Delete subresource when its parent resource is deleted (#326)
* Changed processing order in commonlogk8saudit recorder tasks (#280)

* Changed processing order in commonlogk8saudit recorder tasks

This change enforce the recorder tasks to process parent timelines before its children.
For example, if there were binding subresource under a Pod, then the Pod resource logs are processed at first and then binding resources are processed later to use the parent resource information in its process. This is mostly for handling parent resource deletion properly.
This change is nccessary for fixing bug #274 in following commit.

* fix issues pointed by gemini-code-review

* Remove subresource on parent resource deletion (#281)

* Changed processing order in commonlogk8saudit recorder tasks

This change enforce the recorder tasks to process parent timelines before its children.
For example, if there were binding subresource under a Pod, then the Pod resource logs are processed at first and then binding resources are processed later to use the parent resource information in its process. This is mostly for handling parent resource deletion properly.
This change is nccessary for fixing bug #274 in following commit.

* Define RecorderRequest type to consolidate function arguments in RecoderFunc

Argument count passed to recorder was increased and it's expected to have more fields later.
Defined a type to pass them to the recorder.

* Add TimelineBuilder utility function to retrieve a revision

* fix #274. Delete subresource on parent resource deletion

* fix issues pointed out by gemini-code-assist

GitOrigin-RevId: 1c10204740bffbba6390841aa4c823bfddee0719
Change-Id: I317549067fda88e1a384824a0c28d194cab3bfa1