Added fallbacks when the fields used for main message is not available (#327)
Container log parsers and several parser tasks are depending on the main message field set reader,
but it can show noisy warnings when none of the target log fields are available.

This change fixes the issue by adding fallback fields.
Now it uses fields as a main message in the following order:
 `textPayload` > `jsonPayload.****` (**** would be `message`, `msg`...etc) > jsonPayload > labels

GitOrigin-RevId: e7ee4ecec59a6295fccf8a4c40e1467fa65b6906
Change-Id: I62fb636ccb006585d281912fc903fe30adc9e1bf